+++ bsn1/src/contents_ref.rs
@@ -93,7 +93,7 @@ impl ContentsRef {
         T: PrimInt,
     {
         if self.is_empty() {
-            return Err(Error::UnTerminatedBytes);
+            return Err(Error::UnterminatedBytes);
         }
 
         if 1 < self.len() {
@@ -189,9 +189,9 @@ impl ContentsRef {
     /// ```
     pub const fn to_bool_ber(&self) -> Result<bool, Error> {
         if self.is_empty() {
-            Err(Error::UnTerminatedBytes)
+            Err(Error::UnterminatedBytes)
         } else if 1 < self.len() {
-            Err(Error::InvalidContents)
+            Err(Error::ExtraContentsOctet)
         } else if self.bytes[0] == 0x00 {
             Ok(false)
         } else {
@@ -223,14 +223,14 @@ impl ContentsRef {
     /// ```
     pub const fn to_bool_der(&self) -> Result<bool, Error> {
         if self.is_empty() {
-            Err(Error::UnTerminatedBytes)
+            Err(Error::UnterminatedBytes)
         } else if 1 < self.len() {
-            Err(Error::InvalidContents)
+            Err(Error::ExtraContentsOctet)
         } else {
             match self.bytes[0] {
                 0x00 => Ok(false),
                 0xff => Ok(true),
-                _ => Err(Error::InvalidContents),
+                _ => Err(Error::InvalidDerBooleanContents),
             }
         }
     }